requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
pymongo==4.10.1
sqlalchemy==2.0.40
pydantic==2.12.5
pydantic-settings==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
anthropic==0.89.0
google-genai==1.47.0
openpyxl==3.1.5
python-dotenv==1.2.1
